API/requirements.txt
bidict==0.21.4
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-SocketIO==5.2.0
greenlet==1.1.2
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.0.1
mysql-connector-python==8.0.30
protobuf==3.19.4
psutil==5.9.1
python-engineio==4.3.4
python-socketio==5.7.1
PyYAML==6.0
SQLAlchemy==1.4.40
typing-extensions==4.1.1
Werkzeug==2.0.3
wget==3.2
zipp==3.6.0
